refactor: promote game/utility registries to src/shared/domain (refs #692)
The canonical game list and built-in utility list existed as five parallel
copies that had to be hand-kept in sync: KNOWN_GAME_KEYS + GAMES (game keys),
and KNOWN_UTILITY_KEYS + BUILT_IN_UTILITY_KEYS + BUILT_IN_UTILITIES (utility
keys). Collapse them into one process-agnostic source, src/shared/domain/
registries.ts, with the allowlists derived from GAMES/BUILT_IN_UTILITIES so
they can no longer drift. The old locations re-export from shared, so no
importer changes.

Wiring: add src/shared to the node tsconfig include (now trivial after #752
decoupled the projects) so the main process type-checks its shared import.

Deliberately NOT touched: LEGACY_GAME_KEYS / LEGACY_UTILITY_KEYS in
renderer/src/lib/migrations.ts are a frozen snapshot of the old localStorage
schema (fewer games, no Track Titan, explicit customapp1-5) and must not
follow the live registry, or the one-shot legacy migration would change
behavior as new games are added.

First slice of #692; profile types, the slot scanners (strict + loose pair),
and the small dupes follow; the two normalizers are a separate follow-up.
