+++ src/renderer/src/lib/config.ts
@@ -1,21 +1,15 @@
-export interface Game {
-  key: string
-  name: string
-  icon: string
-}
-export interface Utility {
-  key: string
-  name: string
-  isCustom?: boolean
-  // Bundled curated icon path (mirrors Game.icon). For a built-in slot the
-  // app identity is known, so when this is set the bundled asset is shown
-  // FIRST — ahead of the Windows shell-extracted exe icon — with shell
-  // extraction only as fallback (#727; originally introduced shell-first by
-  // #652). Shell extraction is unreliable across app versions/icon formats
-  // and can "succeed" with a broken image (e.g. Crew Chief's black-square
-  // alpha artifact), which shell-first would keep forever once cached.
-  icon?: string
-}
+// Game/Utility registries live in the process-agnostic domain layer (#692).
+// Re-exported here so the many renderer importers keep their `from '../lib/config'`
+// import path; new code can import from '../../../shared/domain/registries' directly.
+import {
+  GAMES,
+  BUILT_IN_UTILITIES,
+  type Game,
+  type Utility
+} from '../../../shared/domain/registries'
+
+export { GAMES, BUILT_IN_UTILITIES, type Game, type Utility }
+
 export interface ProfileUtility {
   id: string
   enabled: boolean
@@ -68,49 +62,6 @@ export const MAX_CUSTOM_SLOTS = 20
 export function isRecord(value: unknown): value is Record<string, unknown> {
   return !!value && typeof value === 'object' && !Array.isArray(value)
 }
-
-export const GAMES: Game[] = [
-  { key: 'ac', name: 'Assetto Corsa', icon: 'assets/ac.png' },
-  { key: 'acc', name: 'Assetto Corsa Competizione', icon: 'assets/acc.png' },
-  { key: 'acevo', name: 'Assetto Corsa Evo', icon: 'assets/acevo.png' },
-  { key: 'acrally', name: 'Assetto Corsa Rally', icon: 'assets/acrally.png' },
-  { key: 'aeroflyfs4', name: 'Aerofly FS 4', icon: 'assets/aeroflyfs4.png' },
-  { key: 'ams', name: 'Automobilista', icon: 'assets/ams.png' },
-  { key: 'ams2', name: 'Automobilista 2', icon: 'assets/ams2.png' },
-  { key: 'beamng', name: 'BeamNG', icon: 'assets/beamng.png' },
-  { key: 'dcsw', name: 'DCS World', icon: 'assets/dcsw.png' },
-  { key: 'dirtrally', name: 'Dirt Rally', icon: 'assets/dirtrally.png' },
-  { key: 'dirtrally2', name: 'Dirt Rally 2.0', icon: 'assets/dirtrally2.png' },
-  { key: 'eawrc', name: 'EA WRC', icon: 'assets/eawrc.png' },
-  { key: 'f124', name: 'F1 24', icon: 'assets/f124.png' },
-  { key: 'f125', name: 'F1 25', icon: 'assets/f125.png' },
-  { key: 'il2gb', name: 'IL-2 Sturmovik: Great Battles', icon: 'assets/il2gb.png' },
-  { key: 'iracing', name: 'iRacing', icon: 'assets/iracing.png' },
-  { key: 'lmu', name: 'Le Mans Ultimate', icon: 'assets/lmu.png' },
-  { key: 'msfs2020', name: 'Microsoft Flight Simulator 2020', icon: 'assets/msfs2020.png' },
-  { key: 'msfs2024', name: 'Microsoft Flight Simulator 2024', icon: 'assets/msfs2024.png' },
-  { key: 'p3d', name: 'Prepar3D', icon: 'assets/p3d.png' },
-  { key: 'pmr', name: 'Project Motor Racing', icon: 'assets/pmr.png' },
-  { key: 'raceroom', name: 'RaceRoom Racing Experience', icon: 'assets/raceroom.png' },
-  { key: 'rbr', name: 'Richard Burns Rally', icon: 'assets/rbr.png' },
-  { key: 'rennsport', name: 'Rennsport', icon: 'assets/rennsport.png' },
-  { key: 'rf1', name: 'rFactor', icon: 'assets/rf1.png' },
-  { key: 'rf2', name: 'rFactor 2', icon: 'assets/rf2.png' },
-  { key: 'xplane12', name: 'X-Plane 12', icon: 'assets/xplane12.png' }
-]
-
-export const BUILT_IN_UTILITIES: Utility[] = [
-  // Listed first: a telemetry recorder needs to be alive before/at session
-  // start to capture the whole lap, so it defaults to the front of the launch
-  // order (#652).
-  { key: 'tracktitan', name: 'Track Titan', icon: 'assets/tracktitan.png' },
-  { key: 'simhub', name: 'SimHub', icon: 'assets/simhub.png' },
-  { key: 'crewchief', name: 'Crew Chief', icon: 'assets/crewchief.png' },
-  { key: 'tradingpaints', name: 'Trading Paints', icon: 'assets/tradingpaints.png' },
-  { key: 'garage61', name: 'Garage 61', icon: 'assets/garage61.png' },
-  // No bundled asset yet — keeps the shell-icon → initials chain (#727).
-  { key: 'secondmonitor', name: 'Second Monitor' }
-]
 
 export function normalizeCustomSlots(value: unknown): number {
   if (typeof value !== 'number' || !Number.isFinite(value)) {
